server/requirements.txt
aiofiles==23.2.1 ; python_version >= "3.11" and python_version < "4.0"
alembic==1.16.5 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.11.0 ; python_version >= "3.11" and python_version < "4.0"
argon2-cffi-bindings==25.1.0 ; python_version >= "3.11" and python_version < "4.0"
argon2-cffi==23.1.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2025.8.3 ; python_version >= "3.11" and python_version < "4.0"
cffi==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.4.3 ; python_version >= "3.11" and python_version < "4.0"
click==8.3.0 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
cryptography==42.0.8 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.110.3 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.2.4 ; python_version >= "3.11" and python_version < "3.14" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32")
h11==0.16.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.9 ; python_version >= "3.11" and python_version < "4.0"
httptools==0.6.4 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.27.2 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
iniconfig==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
mako==1.3.10 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.11" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.11" and python_version < "4.0"
pillow==11.3.0 ; python_version >= "3.11" and python_version < "4.0"
pluggy==1.6.0 ; python_version >= "3.11" and python_version < "4.0"
pycparser==2.23 ; python_version >= "3.11" and python_version < "4.0" and implementation_name != "PyPy"
pydantic-core==2.33.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.11.9 ; python_version >= "3.11" and python_version < "4.0"
pygments==2.19.2 ; python_version >= "3.11" and python_version < "4.0"
pyjwt==2.10.1 ; python_version >= "3.11" and python_version < "4.0"
pytest-asyncio==0.23.8 ; python_version >= "3.11" and python_version < "4.0"
pytest==8.4.2 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.1.1 ; python_version >= "3.11" and python_version < "4.0"
python-multipart==0.0.6 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.11" and python_version < "4.0"
reportlab==4.4.4 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
sqlalchemy==2.0.43 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.37.2 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.15.0 ; python_version >= "3.11" and python_version < "4.0"
typing-inspection==0.4.1 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.29.0 ; python_version >= "3.11" and python_version < "4.0"
uvloop==0.21.0 ; python_version >= "3.11" and python_version < "4.0" and sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy"
watchfiles==1.1.0 ; python_version >= "3.11" and python_version < "4.0"
websockets==15.0.1 ; python_version >= "3.11" and python_version < "4.0"
